# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ADU Atlas API
  version: "1.0.0"
  summary: >-
    Deterministic, source-cited, address-level preliminary feasibility for
    California ADUs, JADUs, and SB 9 projects.
  description: >
    ADU Atlas is a self-serve, developer-facing California parcel feasibility
    API. Send an address and a proposed project type and receive a fast,
    deterministic, source-cited, timestamped PRELIMINARY feasibility result:
    parcel and zoning context, ADU / JADU / SB 9 rules, setback / height / size
    constraints, hazard and overlay flags, assumptions, confidence, and
    official-source citations.


    The request path is deterministic. It uses versioned structured rules,
    PostGIS spatial joins, and source-linked data only. No large language model
    is used at request time for any legal or geospatial determination. Every
    substantive result field carries provenance (source URL, title, section or
    layer, retrieved and last-verified timestamps, confidence, and data status).


    This API never states that a permit is approved, that a project is legal to
    build, that an outcome is guaranteed, or a final yes/no. It returns a
    feasibility_status of likely_feasible, likely_constrained,
    needs_professional_review, or insufficient_data, and includes a verbatim
    disclaimer in every feasibility response.


    Coverage: Los Angeles City is the v1 target. Other California cities are
...
  contact:
    name: ADU Atlas API Support
...
  license:
...
servers:
...
security:
  - rapidApiKey: []
    rapidApiHost: []
  - directApiKey: []
paths:
  /v1/feasibility:
    post:
      tags: [feasibility]
      operationId: createFeasibilityAnalysis
      summary: Run a preliminary feasibility analysis for one address and project type
      description: >
        Computes a deterministic, source-cited preliminary feasibility result
        for a single address and a single project_type. This is the only
        billable endpoint. A billable unit is one completed analysis that
        resolves to a terminal feasibility_status. Errors, validation failures,
        and unsupported-coverage responses are not billed. Identical inputs from
        the same consumer within 24 hours are served from cache and not
        double-billed. An Idempotency-Key may be supplied to safely retry.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeasibilityRequest"
            examples:
              detachedAduLA:
                summary: Detached ADU in Los Angeles
                value:
                  address: "1234 S Main St, Los Angeles, CA 90015"
                  project_type: detached_adu
                  target_sqft: 800
                  bedrooms: 1
                  proposed_height_ft: 16
                  existing_structure:
                    type: single_family
                    has_garage: true
                  options:
                    near_transit: false
      responses:
        "200":
          description: >-
            Completed feasibility analysis (billable). The feasibility_status is
            terminal and the disclaimer is included verbatim.
          headers:
            Idempotency-Key:
              description: Echo of the supplied idempotency key, if any.
              schema:
                type: string
            X-Billable:
              description: >-
                "true" if this response was metered as a completed analysis,
                "false" if served from the 24h dedupe cache.
              schema:
                type: string
                enum: ["true", "false"]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeasibilityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "422":
          $ref: "#/components/responses/UnsupportedCoverage"
        "429":
          $ref: "#/components/responses/QuotaExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/jurisdictions:
    get:
      tags: [jurisdictions]
      operationId: listJurisdictions
      summary: List jurisdictions and coverage status
      description: >
        Returns every registered jurisdiction with its coverage status,
        supported project types, and the date its sources were last updated.
        Read-only and not billed.
      responses:
        "200":
          description: List of jurisdictions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurisdictionList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/jurisdictions/{slug}/rules:
    get:
      tags: [jurisdictions]
      operationId: getJurisdictionRules
      summary: Get citywide and zone-level rules for a jurisdiction
      description: >
        Returns the citywide and zone-level rule set for a jurisdiction, each
        rule attribute carrying per-field provenance and state-baseline
        compliance flags, plus source citations and version history. Read-only
        and not billed.
      parameters:
        - $ref: "#/components/parameters/JurisdictionSlug"
        - name: zone
          in: query
          required: false
          description: Optional zone code filter (case-insensitive), e.g. R1, RD1.5.
          schema:
            type: string
        - name: project_type
          in: query
          required: false
          description: Optional project type filter.
          schema:
            $ref: "#/components/schemas/ProjectType"
      responses:
        "200":
          description: Rule set for the jurisdiction.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JurisdictionRulesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/analyses/{analysis_id}:
    get:
      tags: [analyses]
      operationId: getAnalysis
      summary: Retrieve a previously computed analysis
      description: >
        Retrieves a stored feasibility analysis by id. Private analyses require
        the same API key or consumer that created them. A public shareable
        analysis may instead be fetched with a share token via the `token` query
        parameter. Read-only and not billed.
      parameters:
        - $ref: "#/components/parameters/AnalysisId"
        - name: token
          in: query
          required: false
          description: >-
            Public share token for a shareable analysis. When supplied, the
            analysis may be retrieved without the originating API key.
          schema:
            type: string
      responses:
        "200":
          description: The stored analysis.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeasibilityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/changelog:
    get:
      tags: [meta]
      operationId: getChangelog
      summary: Public update history by city
      description: >
        Returns the public update history: ingestion runs, rule-version changes,
        and coverage changes, grouped by jurisdiction. Read-only and not billed.
      parameters:
        - name: jurisdiction
          in: query
          required: false
          description: Optional jurisdiction slug filter.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Max entries to return (default 50, max 200).
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Changelog entries, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangelogResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    rapidApiKey:
      type: apiKey
      in: header
      name: X-RapidAPI-Key
      description: >-
        Consumer key injected by the RapidAPI gateway. Present on every request
        proxied through RapidAPI.
    rapidApiHost:
      type: apiKey
      in: header
      name: X-RapidAPI-Host
      description: >-
        Expected gateway host header (aduatlas.p.rapidapi.com). Verified against
        the expected gateway pattern; not used as a secret on its own.
    directApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Direct (non-RapidAPI) API key. The raw key is sha256-hashed server-side
        and matched against the stored key hash. Never send this together with
        RapidAPI gateway headers.

  # -------------------------------------------------------------------------
  # Reusable parameters
  # -------------------------------------------------------------------------
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >-
        Optional client-generated key to make POST /v1/feasibility safe to
        retry. A repeated key with an identical body returns the stored
        response; a repeated key with a different body returns 409.
      schema:
        type: string
        maxLength: 255
    JurisdictionSlug:
      name: slug
      in: path
      required: true
      description: Jurisdiction slug, e.g. los_angeles.
      schema:
        type: string
        examples: [los_angeles]
    AnalysisId:
      name: analysis_id
      in: path
      required: true
      description: Analysis UUID.
      schema:
        type: string
        format: uuid

  # -------------------------------------------------------------------------
  # Reusable responses
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: The request body or parameters were malformed. Not billed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidProjectType:
              value:
                error:
                  code: validation_error
                  message: "project_type must be one of the supported enum values."
                  details:
                    field: project_type
                  request_id: req_01HZY8Q9V0
    Unauthorized:
      description: API credentials missing, invalid, or revoked. Not billed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            missing:
              value:
                error:
                  code: unauthorized
                  message: >-
                    Missing API credentials. Provide RapidAPI gateway headers or
                    the X-API-Key header.
                  request_id: req_01HZY8Q9V1
...
    IdempotencyConflict:
      description: >-
        The Idempotency-Key was reused with a different request body. Not billed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            conflict:
              value:
                error:
                  code: idempotency_key_conflict
                  message: >-
                    This Idempotency-Key was already used with a different
                    request body.
                  request_id: req_01HZY8Q9V4
    UnsupportedCoverage:
      description: >-
        The address resolved to a jurisdiction whose coverage_status is not
        production. No feasibility determination is made and the request is NOT
        billed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            planned:
              value:
                error:
                  code: unsupported_coverage
                  message: >-
                    The city of Oakland is registered but not yet production
                    (coverage_status=planned). No feasibility result is
                    available and you were not billed.
                  details:
                    jurisdiction_slug: oakland
                    coverage_status: planned
                  request_id: req_01HZY8Q9V5
    QuotaExceeded:
      description: Monthly plan quota exceeded. No overages in v1. Not billed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            basic:
              value:
                error:
                  code: quota_exceeded
                  message: >-
                    Monthly quota exceeded for your plan. Upgrade or wait for the
                    next billing cycle. There are no paid overages in v1.
                  details:
                    plan: BASIC
                    monthly_quota: 3
                    used_this_month: 3
                  request_id: req_01HZY8Q9V6
    InternalError:
      description: Unexpected server error. Not billed. No internal detail is leaked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            failure:
              value:
                error:
                  code: internal_error
                  message: "An unexpected error occurred. Please retry."
                  request_id: req_01HZY8Q9V7

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ProjectType:
      type: string
      description: Proposed project type.
      enum:
        - detached_adu
        - attached_adu
        - garage_conversion
        - jadu
        - sb9_duplex
        - sb9_urban_lot_split
...
    FeasibilityRequest:
      type: object
      required: [address, project_type]
      additionalProperties: false
      properties:
        address:
          type: string
          minLength: 5
          description: Free-form street address to analyze.
          examples: ["1234 S Main St, Los Angeles, CA 90015"]
        project_type:
          $ref: "#/components/schemas/ProjectType"
        target_sqft:
          type: [number, "null"]
          minimum: 0
          description: Proposed conditioned floor area in square feet.
        bedrooms:
          type: [integer, "null"]
          minimum: 0
          description: Proposed number of bedrooms.
        proposed_height_ft:
          type: [number, "null"]
          minimum: 0
          description: Proposed structure height in feet.
        existing_structure:
          type: [object, "null"]
          description: Optional context about the existing structure on the parcel.
          properties:
            type:
              type: [string, "null"]
              enum: [single_family, multifamily, none, unknown, null]
            has_garage:
              type: [boolean, "null"]
            year_built:
              type: [integer, "null"]
          additionalProperties: false
        options:
          type: [object, "null"]
          description: Optional flags that influence which state provisions apply.
          properties:
            near_transit:
              type: [boolean, "null"]
              description: Caller-asserted transit proximity hint (verified where possible).
            historic_property:
              type: [boolean, "null"]
            include_envelope:
              type: [boolean, "null"]
              description: Request an approximate conceptual envelope (LA v1 only).
          additionalProperties: false

    # ---- Response building blocks ----
...
    FeasibilityResponse:
      type: object
      required:
        - analysis_id
        - request
        - coverage
        - feasibility_status
        - sources
        - freshness
        - disclaimer
      properties:
        analysis_id:
          type: string
          format: uuid
        request:
          $ref: "#/components/schemas/RequestSummary"
        coverage:
          $ref: "#/components/schemas/CoverageContext"
        parcel:
          $ref: "#/components/schemas/ParcelContext"
        zoning:
          $ref: "#/components/schemas/ZoningContext"
        feasibility_status:
          $ref: "#/components/schemas/FeasibilityStatus"
        score:
          $ref: "#/components/schemas/ScoreBlock"
        eligible_paths:
          type: array
          items:
            $ref: "#/components/schemas/EligiblePath"
        development_constraints:
          $ref: "#/components/schemas/DevelopmentConstraints"
        overlay_findings:
          type: array
          items:
            $ref: "#/components/schemas/OverlayFinding"
        approximate_envelope:
          $ref: "#/components/schemas/ApproximateEnvelope"
        assumptions:
          type: array
          items:
            $ref: "#/components/schemas/Assumption"
        limitations:
          type: array
          items:
            $ref: "#/components/schemas/Limitation"
        sources:
          type: array
          description: All distinct sources used, for top-level citation.
          items:
            $ref: "#/components/schemas/Provenance"
        freshness:
          $ref: "#/components/schemas/Freshness"
        share_token:
          type: [string, "null"]
          description: Present when a public shareable token has been minted.
        disclaimer:
          type: string
          const: >-
            This is preliminary informational zoning and GIS analysis, not
            legal, architectural, surveying, engineering, title, environmental,
            or permit advice. Verify all results with the applicable
            jurisdiction and qualified professionals before making decisions or
            spending money.

    # ---- Jurisdictions ----
...
    JurisdictionList:
      type: object
      required: [data, count]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Jurisdiction"
        count:
          type: integer
          examples: [8]
...